pyproject.toml
[project]
name = "django-img-optimizer"
version = "1.0"
description = "Image optimizer for Django."
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Peter Stavrou"}]
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License"
]
dependencies = ["pillow>=10.2.0"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/peterstavrou/django-img-optimizer"
